Encapsulate list functionality within module
