+++ public/src/main.js
@@ -3,7 +3,7 @@ import { initMap } from './map.js';
 import { newDropdown } from './dropdown.js';
 import { graph } from './graph.js';
 import { candidateGraph } from './candSect.js';
-import { enableSearch } from './search.js';
+import { populateList, initSearch } from './list.js';
 import { enableSort } from './sort.js';
 
 // Leave this blank, value injected during build (see contributing guidelines)
@@ -12,7 +12,7 @@ const MAPBOX_KEY = '';
 // Element functionality initalised here
 function initPage() {
   enableSort();
-  enableSearch();
+  initSearch();
   candidateGraph();
   graph();
   initMap(MAPBOX_KEY);
@@ -33,37 +33,17 @@ function initDropdowns(data) {
 
 // Populates elements whenever a new year's data is retrieved
 function populatePage(data) {
-  // Clear existing rows first
-  const [ clist ] = document.getElementById('candid').getElementsByTagName('tbody');
-  clist.innerHTML = '';
-
-  // Document fragment will trigger reflow only once when attached
-  const newRows = document.createDocumentFragment();
-  data.candidates.forEach(cand => {
-    const row = document.createElement("tr");
-    const name = document.createElement("td");
-    const votes = document.createElement("td");
-
-    name.innerHTML = cand.name;
-    votes.innerHTML = cand.votes;
-
-    row.appendChild(name);
-    row.appendChild(votes);
-    newRows.appendChild(row);
-  });
-
-  // Populate the table with the new data
-  clist.appendChild(newRows);
+  populateList(data);
 }
 
 document.addEventListener('DOMContentLoaded', initPage);
 
 // Immediately send a request off for the initial data needed to populate the page
 getYears().then((data) => {
-  const init = () => {
+  function init() {
     initDropdowns(data);
     populatePage(data);
-  };
+  }
 
   // Populate elements once the DOM is ready (or immediately if already)
   if (document.readyState !== 'loading') {
