--- public/src/search.js
@@ -1,22 +1,0 @@
-function search() {
-  const input = document.getElementById("search");
-  const filter = input.value.toUpperCase();
-  const table = document.getElementById("candid");
-  const tr = table.getElementsByTagName("tr");
-
-  for (const i in tr) {
-    const td = tr[i].getElementsByTagName("td")[0];
-    if (td) {
-      const txtValue = td.textContent || td.innerText;
-      if (txtValue.toUpperCase().indexOf(filter) > -1) {
-        tr[i].style.display = "";
-      } else {
-        tr[i].style.display = "none";
-      }
-    }
-  }
-}
-
-export async function enableSearch() {
-  document.getElementById("search").addEventListener("keyup", search);
-}
